rename : 주문서 컴포넌트를 공용 컴포넌트로 이전, inline_field -> label_value_text 로 컴포넌트 명칭 변경 #32
