+++ app/(main)/order/form/_components/orderer_info_section.tsx
@@ -1,8 +1,8 @@
 "use client";
-import type { TUser } from "@/types";
+import Subheading from "@/components/common/subheading";
+import LabelValueText from "@/components/common/label_value_text";
 
-import Subheading from "@/app/(main)/order/form/_components/subheading";
-import InlineField from "@/app/(main)/order/form/_components/inline_field";
+import type { TUser } from "@/types";
 
 interface OrdererInfoSectionProps {
   user: TUser;
@@ -13,9 +13,9 @@ export default function OrdererInfoSection({ user }: OrdererInfoSectionProps) {
   return (
     <section>
       <Subheading title={"주문자 정보"} />
-      <InlineField label={"주문자"} value={name} />
-      <InlineField label={"휴대폰"} value={phone} />
-      <InlineField label={"이메일"} value={email} />
+      <LabelValueText label={"주문자"} value={name} />
+      <LabelValueText label={"휴대폰"} value={phone} />
+      <LabelValueText label={"이메일"} value={email} />
     </section>
   );
 }
