+++ app/(main)/order/form/_components/payment_info_section.tsx
@@ -1,13 +1,13 @@
 "use client";
 import Button from "@/components/common/button";
+import Subheading from "@/components/common/subheading";
+import LabelValueText from "@/components/common/label_value_text";
+import PaymentButtonList from "@/app/(main)/order/form/_components/payment_button_list";
 
 import styles from "@/app/(main)/order/form/_components/payment_info_section.module.css";
 
 import { PAYMENTS } from "@/constants";
 import classNames from "classnames/bind";
-import Subheading from "@/app/(main)/order/form/_components/subheading";
-import InlineField from "@/app/(main)/order/form/_components/inline_field";
-import PaymentButtonList from "@/app/(main)/order/form/_components/payment_button_list";
 
 interface PaymentInfoSectionProps {
   totalPrice: number;
@@ -20,7 +20,7 @@ export default function PaymentInfoSection({ totalPrice }: PaymentInfoSectionPro
   return (
     <section>
       <Subheading title={"결제 정보"} />
-      <InlineField label={"결제수단 선택"} />
+      <LabelValueText label={"결제수단 선택"} />
       <PaymentButtonList payments={PAYMENTS} />
       <Button
         text={`${formattedTotalPrice}원 결제하기`}
